ok rồi

+++ src/main/java/com/hau/quanlysach/service/NhaXuatBanService.java
@@ -5,11 +5,14 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import com.hau.quanlysach.entity.NhaXuatBan;
 import com.hau.quanlysach.repository.NhaXuatBanRepository;
+import com.hau.quanlysach.repository.SachRepository;
 
 @Service
 public class NhaXuatBanService {
     @Autowired
     private NhaXuatBanRepository repo;
+    @Autowired
+    private SachRepository sachRepo;
 
     public List<NhaXuatBan> getAll() {
         return repo.findAll();
@@ -26,6 +29,12 @@ public class NhaXuatBanService {
     }
 
     public void delete(Integer id) {
+        // Kiểm tra ràng buộc
+        long count = sachRepo.countByNhaXuatBanMaNxb(id);
+        if (count > 0) {
+            throw new RuntimeException(
+                    "Không thể xóa! Nhà xuất bản này đang có " + count + " cuốn sách trong hệ thống.");
+        }
         repo.deleteById(id);
     }
 }
